+++ js-miniapp-sample/src/pages/home.js
@@ -1,4 +1,6 @@
 import React, { useState, useEffect } from 'react';
+import { setQueryParams } from '../services/home/actions';
+import { connect } from 'react-redux';
 
 import {
   makeStyles,
@@ -9,8 +11,13 @@ import {
 } from '@material-ui/core';
 import LogoutIcon from '@mui/icons-material/Logout';
 import clsx from 'clsx';
-import MiniApp from 'js-miniapp-sdk';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import MiniApp, { HostAppEvents } from 'js-miniapp-sdk';
+import {
+  BrowserRouter as Router,
+  Route,
+  Routes,
+  useNavigate,
+} from 'react-router-dom';
 
 import { navItems } from './../routes';
 import Dialogue from '../components/Dialogue';
@@ -33,6 +40,7 @@ const useStyles = makeStyles((theme) => ({
     flexDirection: 'column',
     justifyContent: 'initial',
     alignItems: 'center',
+    overflow: 'auto',
   },
   drawerClosed: {
     width: '100% !important',
@@ -61,9 +69,15 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const Home = (props: any) => {
+type HomeProps = {
+  changeQueryParams: Function,
+};
+
+const Home = (props: HomeProps) => {
   const classes = useStyles();
   const theme = useTheme();
+  const navigate = useNavigate();
+
   const isMobile = useMediaQuery(theme.breakpoints.down('xs'));
   const [shrink, setShrink] = useState(false);
   const [showDrawer, setShowDrawer] = useState(!isMobile);
@@ -72,7 +86,55 @@ const Home = (props: any) => {
 
   useEffect(() => {
     setShowDrawer(!isMobile);
-  }, [isMobile]);
+
+    const onReceivedQueryParams = (e) => {
+      console.log(
+        `on home ${HostAppEvents.DID_RECEIVE_QUERY_PARAMS} Event -> ${e.detail.message}`
+      );
+      let search = new URLSearchParams();
+      try {
+        let url,
+          _message = JSON.parse(e.detail.message);
+        if (e.detail.message.search('navigateTo') === -1) {
+          url = new URL(window.location.toString());
+        }
+        _message.forEach((val, ind, arr) => {
+          if (val.name === 'navigateTo' || val.name === 'scrollTo') {
+            if (!url && val.name === 'navigateTo') {
+              url = new URL(`${window.location.origin}/${val.value}`);
+            }
+            if (val.name === 'scrollTo') {
+              url.hash = val.value;
+            }
+          } else {
+            search.set(val.name, val.value);
+          }
+        });
+        console.log(`URL: ${url.toString()}`);
+        navigate({
+          hash: url.hash,
+          search: search.toString(),
+          pathname: url.pathname,
+        });
+      } catch (e) {
+        console.error(
+          `on home ${HostAppEvents.DID_RECEIVE_QUERY_PARAMS} Error -> ${e}`
+        );
+      }
+      props.changeQueryParams(e.detail.message);
+    };
+
+    window.addEventListener(
+      HostAppEvents.DID_RECEIVE_QUERY_PARAMS,
+      onReceivedQueryParams
+    );
+    return () => {
+      window.removeEventListener(
+        HostAppEvents.DID_RECEIVE_QUERY_PARAMS,
+        onReceivedQueryParams
+      );
+    };
+  }, [navigate, props, isMobile]);
   const onShrinkToggle = (shrinkState) => {
     setShrink(shrinkState);
   };
@@ -101,7 +163,7 @@ const Home = (props: any) => {
     }
   };
   return (
-    <Router>
+    <>
       <ToolBar
         showDrawer={showDrawer}
         onDrawerToggle={onDrawerToggle}
@@ -155,8 +217,28 @@ const Home = (props: any) => {
         contentText="Are you sure you want to logout?"
         errorText={logOutError}
       />
-    </Router>
+    </>
   );
 };
 
-export default Home;
+const mapStateToProps = (state, props) => {
+  return {
+    ...props,
+  };
+};
+
+const mapDispatchToProps = (dispatch) => {
+  return {
+    changeQueryParams: (payload) => dispatch(setQueryParams(payload)),
+  };
+};
+
+const ConnectedHome = connect(mapStateToProps, mapDispatchToProps)(Home);
+
+const HomePage = () => (
+  <Router>
+    <ConnectedHome />
+  </Router>
+);
+
+export default HomePage;