+++ js-miniapp-sdk/CHANGELOG.md
@@ -1,7 +1,13 @@
 ## CHANGELOG
 
-### 1.25.0 (2025-XX-XX)
+### 1.26.0 (2025-09-01)
+- **Feature:** Added new interfaces `launchAppUsingDeeplink` and `launchAppUsingPackageName` to MiniAppUtils for launching external apps via deeplink or package name.
+- **Feature:** Added support with `HostAppEvents` to receive any Query Parameters as JSON/String from the HostApp to MiniApp with event `DID_RECEIVE_QUERY_PARAMS`.
+
+### 1.25.0 (2025-06-09)
 - **Feature:** `launchInternalBrowser` now supports passing an options object (`LaunchBrowserOptions`) with `httpMethod`, `httpBody`, `audience`, and `scopes`.
+- **Feature:** Added new interface `getPermissionStatus` to check the status of a permission.
+- **Feature:** Added new interface `launchAppSettings` to open the device's app settings screen.
 ---
 
 ### 1.24.0 (2025-04-07)